more warning suppression

+++ pandas/_libs/src/vendored/ujson/python/objToJSON.c
@@ -1491,10 +1491,13 @@ static void Object_beginTypeContext(JSOBJ _obj, JSONTypeContext *tc) {
     }
     return;
   } else if (PyDelta_Check(obj)) {
-    npy_int64 value =
-        PyObject_HasAttrString(obj, "_value") ? get_long_attr(obj, "_value")
-                                              : // pd.Timedelta object or pd.NaT
-            total_seconds(obj) * 1000000000LL;  // nanoseconds per sec
+    // pd.Timedelta object or pd.NaT should evaluate true here
+    // fallback to nanoseconds per sec for other objects
+    // TODO(anyone): cast below loses precision if total_seconds return
+    // value exceeds number of bits that significand can hold
+    npy_int64 value = PyObject_HasAttrString(obj, "_value")
+                          ? get_long_attr(obj, "_value")
+                          : (int64_t)total_seconds(obj) * 1000000000LL;
 
     if (value == get_nat()) {
       tc->type = JT_NULL;
